requirements/base.txt
#DJANGO
Django~=4.1.5
psycopg2-binary~=2.9.9
django-environ~=0.9.0
djangorestframework~=3.14.0
drf-spectacular~=0.27.0
drf-spectacular-sidecar~=2023.9.1
uvicorn[standard]~=0.20.0
gunicorn~=20.1.0
django-phonenumber-field~=7.3.0
phonenumbers~=8.13.29
djangorestframework-simplejwt~=5.3.1
smsir-python~=1.0.7
requests~=2.31.0
celery~=5.3.6
redis~=5.0.1
pillow~=10.3.0
python-magic~=0.4.27

